# app/definition/swagger.yaml
swagger: "2.0"
info:
  version: desp - v1
  title: Despachalo.com
  description: API de servicios de Despachalo.com
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /desp/v1
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /userLogin:
    post:
      tags:
        - Usuario
      description: Recibe el correo y contraseña del usuario para iniciar sesión en el sistema
      # used as the method name of the controller
      operationId: userLogin
      parameters:
        - name: data
          in: body
          required: true
          schema:
            $ref: "#/definitions/UserLogin"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DespachaloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users:
    get:
      tags:
        - Usuario
      description: Lista todos los usuarios
      # used as the method name of the controller
      operationId: findUsers
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DespachaloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      tags:
        - Usuario
      description: Obtener información de usuario
      # used as the method name of the controller
      operationId: findUserByEmail
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: data
          in: body
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DespachaloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      tags:
        - Usuario
      description: Editar usuario
      # used as the method name of the controller
      operationId: updateUser
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: data
          in: body
          required: true
          schema:
            $ref: "#/definitions/UserEdit"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DespachaloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      tags:
        - Usuario
      description: Cambia el estado del usuario a Inactivo
      # used as the method name of the controller
      operationId: deleteUser
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: data
          in: body
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DespachaloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users/types:
    get:
      tags:
        - Usuario
      description: Obtener los tipos de usuario
      # used as the method name of the controller
      operationId: findTypesUser
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DespachaloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      tags:
        - Usuario
      description: Obtener el tipo de usuario
      # used as the method name of the controller
      operationId: findTypeUser
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: data
          in: body
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DespachaloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users/add:
    post:
      tags:
        - Usuario
      description: Registrar usuarios de tipo Administrador
      # used as the method name of the controller
      operationId: addUser
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: data
          in: body
          required: true
          schema:
            $ref: "#/definitions/UserAdd"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DespachaloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users/resetPassword:
    get:
      tags:
        - Usuario
      description: Reestablece la contraseña de un usuario Admin
      # used as the method name of the controller
      operationId: userResetPassword
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: admin
          in: query
          description: Correo del usuario administrador
          required: true
          type: string
        - name: email
          in: query
          description: Correo del usuario
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DespachaloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users/register:
    post:
      tags:
        - Usuario
      description: Registrar usuarios clientes
      # used as the method name of the controller
      operationId: userRegister
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: data
          in: body
          required: true
          schema:
            $ref: "#/definitions/UserRegister"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DespachaloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users/validateEmail:
    get:
      tags:
        - Usuario
      description: Validar email de usuario
      # used as the method name of the controller
      operationId: userValidateEmail
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: email
          in: query
          description: Correo del usuario
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DespachaloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  User:
    type: object
    properties:
      email:
        description: Correo del usuario
        type: string
  UserLogin:
    type: object
    required:
      - email
      - password
    properties:
        email:
          description: Correo del usuario
          type: string
        password:
          description: Contraseña del usuario
          type: string
          format: password
  UserRegister:
    type: object
    required:
      - email
    properties:
      email:
        description: Correo del usuario
        type: string
      name:
        description: Nombre del usuario
        type: string
      lastname:
        description: Apellido del usuario
        type: string
      type:
        description: Tipo del usuario
        type: string
  UserAdd:
    type: object
    required:
      - email
      - name
      - lastname
      - type
      - rut
    properties:
      email:
        description: Correo del usuario
        type: string
      name:
        description: Nombre del usuario
        type: string
      lastname:
        description: Apellido del usuario
        type: string
      type:
        description: Tipo3 del usuario
        type: string
      rut:
        description: Rut del usuario
        type: string
  UserEdit:
    type: object
    required:
      - email
    properties:
      email:
        description: Correo del usuario
        type: string
      name:
        description: Nombre del usuario
        type: string
      lastname:
        description: Apellido del usuario
        type: string
      password:
        description: Contraseña del usuario
        type: string
        format: password
      rut:
        description: Rut del usuario
        type: string
      type:
        description: Tipo del usuario
        type: string
...
  DespachaloResponse:
    type: string
  ErrorResponse:
    type: string